requirements.txt
Flask==3.0.0
gunicorn==23.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
markdown==3.5.1
